[FIX] delivery_sendcloud: batch parcels with float avg weight
Steps to reproduce:

- Configure the sendcloud delivery method (for instance with Bpost @home
  for a belgian company) and enable the `Use Batch Shipping` option
  on the delivery method.
- Create a storable product with a weight of 1 kg and a positive volume.
- Create and confirm sale order for 4 units
- Add shipping -> chose sendcloud
- Separate the delivery of 4 kg in 3 packs:
    - set quantity to 1 -> put in pack
    - set quantity to 2 -> put in pack
    - set the quantity to 4 -> put in pack
- Try to validate the delivery
> invalid operation:  weight: "A valid integer is required."

Cause of the issue:

Sendcloud's api only accept integer values for the weight of a product.
However, to evaluate the price of the parcel accurately for a batch
shipping we need to compute the average weigth to provide to sendcloud.
Converted to grams we therefore provide a value of 1333.333333 to
sendcloud which raises an invalid operation:
https://github.com/odoo/enterprise/blob/72b4a223ec4e2e09fc93bebd150923abba37a8df/delivery_sendcloud/models/sendcloud_service.py#L378-L379
https://github.com/odoo/enterprise/blob/72b4a223ec4e2e09fc93bebd150923abba37a8df/delivery_sendcloud/models/sendcloud_service.py#L205-L207
https://github.com/odoo/enterprise/blob/72b4a223ec4e2e09fc93bebd150923abba37a8df/delivery_sendcloud/models/sendcloud_service.py#L44

opw-4874063

closes odoo/enterprise#89686

X-original-commit: 04b71e19a54a333127138c88c34c4ba052108c99

+++ delivery_sendcloud/models/sendcloud_service.py
@@ -387,7 +387,7 @@ class SendCloud:
         use_multicollo = carrier_id.sendcloud_use_batch_shipping and to_europe
         single_shipping = len(delivery_packages) == 1 or (use_multicollo and len(delivery_packages) <= 20)
         #Avg weight for multiple packages in single shipping, sendcloud multiplies it with number of packages
-        api_weight = shipping_weight/len(delivery_packages) if single_shipping else None
+        api_weight = int(shipping_weight / len(delivery_packages)) if single_shipping else None
 
         # Fetch shipping methods compatible with current picking
         shipping_methods = self._get_shipping_methods(picking.carrier_id, from_country, to_country, api_weight, is_return, from_postal_code=from_postal_code, to_postal_code=to_postal_code)
